refactor: S3Folder, ImageUploader 위치 변경
