+++ src/main/java/com/spaceclub/club/service/ClubService.java
@@ -5,8 +5,8 @@ import com.spaceclub.club.domain.ClubUser;
 import com.spaceclub.club.repository.ClubRepository;
 import com.spaceclub.club.repository.ClubUserRepository;
 import com.spaceclub.club.service.vo.ClubInfo;
-import com.spaceclub.global.config.s3.S3Folder;
-import com.spaceclub.global.config.s3.S3ImageUploader;
+import com.spaceclub.global.s3.S3Folder;
+import com.spaceclub.global.s3.S3ImageUploader;
 import com.spaceclub.global.config.s3.S3Properties;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
