+++ src/main/java/com/spaceclub/event/service/EventService.java
@@ -10,8 +10,8 @@ import com.spaceclub.event.repository.EventUserRepository;
 import com.spaceclub.event.service.vo.ClubEventOverviewGetInfo;
 import com.spaceclub.event.service.vo.EventCreateInfo;
 import com.spaceclub.event.service.vo.UserBookmarkedEventGetInfo;
-import com.spaceclub.global.config.s3.S3Folder;
-import com.spaceclub.global.config.s3.S3ImageUploader;
+import com.spaceclub.global.s3.S3Folder;
+import com.spaceclub.global.s3.S3ImageUploader;
 import com.spaceclub.global.config.s3.S3Properties;
 import com.spaceclub.user.service.UserProvider;
 import com.spaceclub.user.service.vo.UserProfile;
