+++ src/main/java/com/spaceclub/global/s3/S3Folder.java
@@ -1,4 +1,4 @@
-package com.spaceclub.global.config.s3;
+package com.spaceclub.global.s3;
 
 import lombok.Getter;
